Add gap-filling task
Oops, forgot to commit for a while.

Now we have a whole new paradigm + experiment that takes the kinds of templates I was producing over
the QAMR data and crowdsources new answers to questions.

All the basic functionality (including question rendering) is there, and now I just need to improve
question suggestion. Possible next to-dos are:

 * add verb-local count information to improve the suggester's accuracy
 * include questions we already have in the data into suggestion to avoid redundancy
 * add paraphrasing judgments to the suggester to reduce redundancy
 * incorporate non-fully-aligned questions (i.e. with "something") to improve paraphrasing judgments
 * Develop a way of compressing/pruning the counts data structure to send to the client